Able to add user to Mongo DB and refresh on app but userInfo redux state is large (the description is listed in the description))
utils.ts:18 SerializableStateInvariantMiddleware took 136ms, which is more than the warning threshold of 32ms.
If your state or actions are very large, you may want to disable the middleware as it might cause too much of a slowdown in development mode. See https://redux-toolkit.js.org/api/getDefaultMiddleware for instructions.
It is disabled in production builds, so you don't need to worry about that.


Former-commit-id: 66397429d2aed99aa1a022508adc197ab57e192d